added apiService.js to handle the getting of the data, added apiKeys.js to save the api keys in a secure file
