+++ src/apiKeys.js
@@ -1,0 +1,2 @@
+export const GOOGLE_API_KEY = "AIzaSyAyeqryVcCKYGXm0miOQuRsyifHTk2I-Fo";
+export const OPENWEATHER_API_KEY = "0fb9ebda5842e2a12a43af55c70d2e3f";
